requirements.txt
aiofiles==22.1.0
aioredis==1.3.1
aligo==5.4.0
amqp==2.6.1
asgiref==3.5.2
async-timeout==4.0.2
backoff==2.0.1
bcrypt==3.2.0
billiard==3.6.4.0
cachetools==4.2.2
celery==4.4.0
celery-pool-asyncio==0.1.11
Cerberus==1.3.3
certifi==2022.9.24
cffi==1.15.0
charset-normalizer==2.1.1
ci-info==0.3.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coloredlogs==15.0.1
configobj==5.0.6
configparser==5.3.0
cryptography==38.0.3
etelemetry==0.3.0
filelock==3.8.0
fitz==0.0.1.dev2
future==0.18.2
greenlet==2.0.1
hiredis==2.0.0
httplib2==0.21.0
httptools==0.5.0
humanfriendly==10.0
idna==3.4
isodate==0.6.1
kombu==4.6.11
looseversion==1.0.2
lxml==4.9.1
msgpack==1.0.4
multidict==5.2.0
networkx==2.8.8
nibabel==4.0.2
nipype==1.8.5
numpy==1.23.4
packaging==21.3
pandas==1.5.2
pathlib==1.0.1
Pillow==9.3.0
prompt-toolkit==3.0.32
prov==2.0.0
pycparser==2.21
pycryptodome==3.15.0
pycryptodomex==3.9.4
pydot==1.4.2
PyMuPDF==1.21.0
PyMySQL==0.9.3
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-alipay-sdk==3.0.4
python-dateutil==2.8.2
pytz==2022.6
pyxnat==1.5
PyYAML==6.0
qiniu==7.7.1
qrcode==7.3.1
qrcode-terminal==0.8
rdflib==6.2.0
redis==3.5.3
requests==2.28.1
sanic==21.9.3
sanic-openapi==21.6.1
sanic-routing==0.7.2
scipy==1.9.3
simplejson==3.17.6
six==1.16.0
SQLAlchemy==1.4.25
tqdm==4.64.1
traits==6.3.2
typing_extensions==4.4.0
ujson==4.2.0
urllib3==1.26.12
uvloop==0.17.0
vine==1.3.0
wcwidth==0.2.5
websockets==10.4
